Add handling for spark correlations with no numeric fields
This change addresses issue #1722 (https://github.com/ydataai/ydata-profiling/issues/1722).

Assembling a vector column in Spark with no numeric columns results in features with a NULL size, NULL indices, and an empty list of values.
This causes an exception to be raised when computing correlations.

The solution here is to avoid computing the correlation matrix when there are no interval columns (numeric).